Huge upstream merge

+++ code/game/objects/structures/xeno.dm
@@ -14,11 +14,24 @@
 	var/on_fire = FALSE
 	///Set this to true if this object isn't destroyed when the weeds under it is.
 	var/ignore_weed_destruction = FALSE
+	///Which hive it belongs to
+	var/hivenumber = XENO_HIVE_NORMAL
 
-/obj/alien/Initialize(mapload)
+/obj/alien/Initialize(mapload, _hivenumber)
 	. = ..()
+	if(_hivenumber) ///because admins can spawn them
+		hivenumber = _hivenumber
+	var/datum/hive_status/hive = GLOB.hive_datums[hivenumber]
+	name = "[hive.prefix][name]"
+	color = hive.color
 	if(!ignore_weed_destruction)
 		RegisterSignal(loc, COMSIG_TURF_WEED_REMOVED, PROC_REF(weed_removed))
+
+/obj/alien/update_icon_state()
+	.=..()
+	var/datum/hive_status/hive = GLOB.hive_datums[hivenumber]
+	if(istype(hive))
+		color = hive.color
 
 /// Destroy the alien effect when the weed it was on is destroyed
 /obj/alien/proc/weed_removed()
@@ -69,7 +82,7 @@
 
 
 /obj/alien/resin/attack_hand(mob/living/user)
-	balloon_alert(user, "You only scrape at it")
+	balloon_alert(user, "need a weapon!")
 	return TRUE
 
 
@@ -80,7 +93,8 @@
 	density = FALSE
 	opacity = FALSE
 	max_integrity = 36
-	layer = BELOW_OBJ_LAYER
+	plane = FLOOR_PLANE
+	layer = ABOVE_WEEDS_LAYER
 	hit_sound = SFX_ALIEN_RESIN_MOVE
 	var/slow_amt = 8
 	/// Does this refund build points when destoryed?
@@ -105,7 +119,18 @@
 		COOLDOWN_INCREMENT(vehicle, cooldown_vehicle_move, WEED_SLOWDOWN)
 		return
 
-	if(!ishuman(crosser))
+	if(isxeno(crosser))
+		var/mob/living/carbon/xenomorph/X = crosser
+		if(!issamexenohive(X))
+			X.next_move_slowdown += slow_amt
+		else
+			X.next_move_slowdown += X.xeno_caste.weeds_speed_mod
+		return
+
+	if(issamexenohive(crosser))
+		return
+
+	if(!isliving(crosser))
 		return
 
 	if(HAS_TRAIT(crosser, TRAIT_TANK_DESANT))
@@ -114,26 +139,27 @@
 	if(CHECK_MULTIPLE_BITFIELDS(crosser.allow_pass_flags, HOVERING))
 		return
 
-	var/mob/living/carbon/human/victim = crosser
+	var/mob/living/victim = crosser
 
 	if(victim.lying_angle)
 		return
 
 	victim.next_move_slowdown += slow_amt
 
-/obj/alien/resin/sticky/attack_alien(mob/living/carbon/xenomorph/xeno_attacker, damage_amount = xeno_attacker.xeno_caste.melee_damage, damage_type = BRUTE, armor_type = MELEE, effects = TRUE, armor_penetration = xeno_attacker.xeno_caste.melee_ap, isrightclick = FALSE)
+/obj/alien/resin/sticky/attack_alien(mob/living/carbon/xenomorph/xeno_attacker, damage_amount = xeno_attacker.xeno_caste.melee_damage * xeno_attacker.xeno_melee_damage_modifier, damage_type = BRUTE, armor_type = MELEE, effects = TRUE, armor_penetration = xeno_attacker.xeno_caste.melee_ap, isrightclick = FALSE)
 	if(xeno_attacker.status_flags & INCORPOREAL)
 		return FALSE
 
+	if(!(issamexenohive(xeno_attacker)))
+		return ..()
+
 	if(xeno_attacker.a_intent == INTENT_HARM) //Clear it out on hit; no need to double tap.
 		if(CHECK_BITFIELD(SSticker.mode?.round_type_flags, MODE_ALLOW_XENO_QUICKBUILD) && SSresinshaping.active && refundable)
-			SSresinshaping.quickbuild_points_by_hive[xeno_attacker.hivenumber]++
+			SSresinshaping.quickbuild_points_by_hive[xeno_attacker.get_xeno_hivenumber()]++
 		xeno_attacker.do_attack_animation(src, ATTACK_EFFECT_CLAW) //SFX
 		playsound(src, SFX_ALIEN_RESIN_BREAK, 25) //SFX
 		deconstruct(TRUE)
 		return
-
-	return ..()
 
 /obj/alien/resin/sticky/can_z_move(direction, turf/start, turf/destination, z_move_flags, mob/living/rider)
 	z_move_flags |= ZMOVE_ALLOW_ANCHORED
@@ -159,7 +185,7 @@
 	icon = 'icons/obj/smooth_objects/resin-door.dmi'
 	icon_state = "resin-door-1"
 	base_icon_state = "resin-door"
-	resistance_flags = NONE
+	resistance_flags = UNACIDABLE | XENO_DAMAGEABLE
 	layer = BELOW_OBJ_LAYER
 	max_integrity = 100
 	smoothing_flags = SMOOTH_BITMASK
@@ -178,27 +204,31 @@
 	var/close_delay = 10 SECONDS
 	///The timer that tracks the delay above
 	var/closetimer
+	var/hivenumber = XENO_HIVE_NORMAL
 
 /obj/structure/mineral_door/resin/smooth_icon()
 	. = ..()
 	update_icon()
 
-/obj/structure/mineral_door/resin/Initialize(mapload)
+/obj/structure/mineral_door/resin/Initialize(mapload, _hivenumber)
 	. = ..()
+	if(_hivenumber) ///because admins can spawn them
+		hivenumber = _hivenumber
+	var/datum/hive_status/hive = GLOB.hive_datums[hivenumber]
+	name = "[hive.prefix][name]"
+	color = hive.color
 	if(!locate(/obj/alien/weeds) in loc)
 		new /obj/alien/weeds(loc)
 
 
 /obj/structure/mineral_door/resin/Cross(atom/movable/mover, turf/target)
 	. = ..()
-	if(!. && isxeno(mover) && !open)
+	if(!. && issamexenohive(mover) && !open)
+		var/mob/living/living_mover = mover
+		if((!(istype(living_mover))) || (living_mover.stat != CONSCIOUS))
+			return
 		toggle_state()
 		return TRUE
-	if(ishuman(mover))
-		var/mob/living/carbon/human/H = mover
-		if(!. && H.faction == FACTION_CLF)
-			if(!open)
-				return TRUE
 
 /obj/structure/mineral_door/resin/attack_larva(mob/living/carbon/xenomorph/larva/M)
 	var/turf/cur_loc = M.loc
@@ -208,23 +238,30 @@
 	return TRUE
 
 //clicking on resin doors attacks them, or opens them without harm intent
-/obj/structure/mineral_door/resin/attack_alien(mob/living/carbon/xenomorph/xeno_attacker, damage_amount = xeno_attacker.xeno_caste.melee_damage, damage_type = BRUTE, armor_type = MELEE, effects = TRUE, armor_penetration = xeno_attacker.xeno_caste.melee_ap, isrightclick = FALSE)
+/obj/structure/mineral_door/resin/attack_alien(mob/living/carbon/xenomorph/xeno_attacker, damage_amount = xeno_attacker.xeno_caste.melee_damage * xeno_attacker.xeno_melee_damage_modifier, damage_type = BRUTE, armor_type = MELEE, effects = TRUE, armor_penetration = xeno_attacker.xeno_caste.melee_ap, isrightclick = FALSE)
 	var/turf/cur_loc = xeno_attacker.loc
 	if(!istype(cur_loc))
 		return FALSE //Some basic logic here
+	if(!issamexenohive(xeno_attacker))
+		return ..()
 	if(xeno_attacker.a_intent != INTENT_HARM)
 		try_toggle_state(xeno_attacker)
 		return TRUE
 	if(CHECK_BITFIELD(SSticker.mode?.round_type_flags, MODE_ALLOW_XENO_QUICKBUILD) && SSresinshaping.active)
-		SSresinshaping.quickbuild_points_by_hive[xeno_attacker.hivenumber]++
+		SSresinshaping.quickbuild_points_by_hive[xeno_attacker.get_xeno_hivenumber()]++
 		qdel(src)
 		return TRUE
 
-	src.balloon_alert(xeno_attacker, "Destroying...")
+	src.balloon_alert(xeno_attacker, "destroying...")
 	playsound(src, SFX_ALIEN_RESIN_BREAK, 25)
 	if(do_after(xeno_attacker, 1 SECONDS, IGNORE_HELD_ITEM, src, BUSY_ICON_HOSTILE))
-		src.balloon_alert(xeno_attacker, "Destroyed")
+		src.balloon_alert(xeno_attacker, "destroyed")
 		qdel(src)
+
+/obj/structure/mineral_door/resin/take_damage(damage_amount, damage_type, armor_type, effects, attack_dir, armour_penetration, mob/living/blame_mob)
+	if(damage_type != BRUTE && damage_type != BURN)
+		return
+	return ..()
 
 /obj/structure/mineral_door/resin/fire_act(burn_level)
 	take_damage(burn_level * 2, BURN, FIRE)
@@ -241,7 +278,7 @@
 			take_damage(30, BRUTE, BOMB)
 
 /obj/structure/mineral_door/resin/try_toggle_state(atom/user)
-	if(isxeno(user))
+	if(issamexenohive(user))
 		return ..()
 
 /obj/structure/mineral_door/resin/toggle_state()
@@ -285,7 +322,7 @@
 			. = TRUE
 			break
 	if(!.)
-		src.balloon_alert_to_viewers("Collapsed")
+		src.balloon_alert_to_viewers("collapsed")
 		qdel(src)
 
 /obj/structure/mineral_door/resin/thick
@@ -297,24 +334,38 @@
 	icon = 'icons/Xeno/xeno_materials.dmi'
 	icon_state = "resin_jelly"
 	soft_armor = list(MELEE = 0, BULLET = 0, LASER = 0, ENERGY = 0, BOMB = 0, BIO = 0, FIRE = 200, ACID = 0)
+	/// The amount of deciseconds to give the Resin Coating status effect.
 	var/immune_time = 30 SECONDS
-	///Holder to ensure only one user per resin jelly.
+	/// Holder to ensure only one user per resin jelly.
 	var/current_user
+	/// If thrown, should it create a 3x3 when it lands? If so, how long in deciseconds if it impacts a human?
+	var/combustive_duration = 0 SECONDS
+	var/hivenumber = XENO_HIVE_NORMAL
 
-/obj/item/resin_jelly/attack_alien(mob/living/carbon/xenomorph/xeno_attacker, damage_amount = xeno_attacker.xeno_caste.melee_damage, damage_type = BRUTE, armor_type = MELEE, effects = TRUE, armor_penetration = xeno_attacker.xeno_caste.melee_ap, isrightclick = FALSE)
+/obj/item/resin_jelly/Initialize(mapload, _hivenumber)
+	. = ..()
+	if(_hivenumber) ///because admins can spawn them
+		hivenumber = _hivenumber
+	var/datum/hive_status/hive = GLOB.hive_datums[hivenumber]
+	name = "[hive.prefix][name]"
+	color = hive.color
+
+/obj/item/resin_jelly/attack_alien(mob/living/carbon/xenomorph/xeno_attacker, damage_amount = xeno_attacker.xeno_caste.melee_damage * xeno_attacker.xeno_melee_damage_modifier, damage_type = BRUTE, armor_type = MELEE, effects = TRUE, armor_penetration = xeno_attacker.xeno_caste.melee_ap, isrightclick = FALSE)
 	if(xeno_attacker.status_flags & INCORPOREAL)
 		return FALSE
-
 	return attack_hand(xeno_attacker)
 
 /obj/item/resin_jelly/attack_self(mob/living/carbon/xenomorph/user)
 	//Activates if the item itself is clicked in hand.
 	if(!isxeno(user))
 		return
+	if(!issamexenohive(user))
+		user.balloon_alert(user, "Wrong hive")
+		return
 	if(user.do_actions || !isnull(current_user))
 		return
 	current_user = user
-	user.balloon_alert(user, "Applying...")
+	user.balloon_alert(user, "applying...")
 	if(!do_after(user, RESIN_SELF_TIME, TRUE, user, BUSY_ICON_MEDICAL))
 		current_user = null
 		return
@@ -326,14 +377,17 @@
 	if(!isxeno(user))
 		return TRUE
 	if(!isxeno(M))
-		M.balloon_alert(user, "Cannot apply")
+		M.balloon_alert(user, "that's not a xeno!")
 		return FALSE
 	if(user.do_actions || !isnull(current_user))
 		return FALSE
+	if(!issamexenohive(user))
+		user.balloon_alert(user, "Wrong hive")
+		return FALSE
 	current_user = M
-	M.balloon_alert(user, "Applying...")
+	M.balloon_alert(user, "applying...")
 	if(M != user)
-		user.balloon_alert(M, "Applying jelly...") //Notify recipient to not move.
+		user.balloon_alert(M, "applying jelly...") //Notify recipient to not move.
 	if(!do_after(user, (M == user ? RESIN_SELF_TIME : RESIN_OTHER_TIME), TRUE, M, BUSY_ICON_MEDICAL))
 		current_user = null
 		return FALSE
@@ -350,19 +404,49 @@
 
 /obj/item/resin_jelly/throw_at(atom/target, range, speed, thrower, spin, flying = FALSE, targetted_throw = TRUE)
 	if(isxenohivelord(thrower))
-		RegisterSignal(src, COMSIG_MOVABLE_IMPACT, PROC_REF(jelly_throw_hit))
+		RegisterSignal(src, COMSIG_MOVABLE_IMPACT, PROC_REF(on_throw_impact))
+		RegisterSignal(src, COMSIG_MOVABLE_POST_THROW, PROC_REF(on_throw_ended))
 	. = ..()
 
-/obj/item/resin_jelly/proc/jelly_throw_hit(datum/source, atom/hit_atom)
+/// Applies the jelly effect for xenomorphs. Otherwise, may explode into sticky resin against everyone else.
+/obj/item/resin_jelly/proc/on_throw_impact(datum/source, atom/hit_atom, speed)
 	SIGNAL_HANDLER
-	UnregisterSignal(source, COMSIG_MOVABLE_IMPACT)
+	if(!isliving(hit_atom))
+		return
+	UnregisterSignal(source, list(COMSIG_MOVABLE_IMPACT, COMSIG_MOVABLE_POST_THROW))
+	var/mob/living/hit_living = hit_atom
+	if(combustive_duration)
+		for(var/turf/sticky_tile AS in RANGE_TURFS(1, get_turf(hit_living)))
+			if(!locate(/obj/alien/resin/sticky/thin) in sticky_tile.contents)
+				var/obj/alien/resin/sticky/thin/temporary_resin = new(sticky_tile)
+				QDEL_IN(temporary_resin, 15 SECONDS)
+		if(!hit_living.issamexenohive(thrower))
+			hit_living.adjust_stagger(combustive_duration)
+		playsound(loc, SFX_ALIEN_RESIN_BUILD, 50, 1)
+		qdel(src)
+		return
 	if(!isxeno(hit_atom))
 		return
-	var/mob/living/carbon/xenomorph/X = hit_atom
-	if(X.xeno_caste.caste_flags & CASTE_FIRE_IMMUNE)
+	var/mob/living/carbon/xenomorph/hit_xenomorph = hit_atom
+	if(hit_xenomorph.xeno_caste.caste_flags & CASTE_FIRE_IMMUNE)
 		return
-	X.visible_message(span_notice("[X] is splattered with jelly!"))
-	INVOKE_ASYNC(src, PROC_REF(activate_jelly), X)
+	if(!issamexenohive(hit_xenomorph))
+		return
+	hit_xenomorph.visible_message(span_notice("[hit_xenomorph] is splattered with jelly!"))
+	INVOKE_ASYNC(src, PROC_REF(activate_jelly), hit_xenomorph)
+
+/// Possibly explode into sticky resin upon finishing the throw.
+/obj/item/resin_jelly/proc/on_throw_ended(datum/source)
+	SIGNAL_HANDLER
+	UnregisterSignal(source, list(COMSIG_MOVABLE_IMPACT, COMSIG_MOVABLE_POST_THROW))
+	if(!combustive_duration)
+		return
+	for(var/turf/sticky_tile AS in RANGE_TURFS(1, loc))
+		if(!locate(/obj/alien/resin/sticky/thin) in sticky_tile.contents)
+			var/obj/alien/resin/sticky/thin/temporary_resin = new(sticky_tile, hivenumber)
+			QDEL_IN(temporary_resin, 15 SECONDS)
+	playsound(loc, SFX_ALIEN_RESIN_BUILD, 50, 1)
+	qdel(src)
 
 ///////////////////////
 /// Globadier Mines ///
@@ -508,4 +592,33 @@
 	var/datum/effect_system/smoke_spread/xeno/hemodile/gas = new(get_turf(src))
 	gas.set_up(2, src)
 	gas.start()
+	qdel(src)
+
+/obj/item/reagent_containers/food/snacks/nutrient_jelly
+	name = "nutrient jelly"
+	desc = "A perplexing, soft mesh of almost bready protein-fibers. It's warm and spongey to the touch, and smells edible."
+	icon = 'icons/obj/items/food/xeno.dmi'
+	icon_state = "edible_biomass_mono"
+	greyscale_config = /datum/greyscale_config/nutrient_jelly
+	greyscale_colors = "#f47e7e"
+	bitesize = 3
+	list_reagents = list(/datum/reagent/consumable/nutriment = 2, /datum/reagent/consumable/nutriment/protein = 5, /datum/reagent/consumable/nutriment/vitamin = 1)
+	tastes = list("vague sweetness" = 1, "water" = 2)
+	var/hivenumber = XENO_HIVE_NORMAL
+
+/obj/item/reagent_containers/food/snacks/nutrient_jelly/examine(mob/user)
+	. = ..()
+	var/datum/hive_status/hive = GLOB.hive_datums[hivenumber]
+	. += "It is from the [hive.name] hive"
+
+/obj/item/reagent_containers/food/snacks/nutrient_jelly/attack_self(mob/living/carbon/xenomorph/user)
+	//Activates if the item itself is clicked in hand.
+	if(!isxeno(user))
+		return
+	if(!issamexenohive(user))
+		user.balloon_alert(user, "Wrong hive")
+		return
+	user.balloon_alert(user, "Absorbing...")
+	if(!do_after(user, RESIN_SELF_TIME, TRUE, user, BUSY_ICON_MEDICAL))
+		return
 	qdel(src)
